Adding support for Polylang "Hide URL language"
`MultilingualPolylang::get_permalink()` does not yet support the option "Hide URL language information for default language". If set true in the Polylang settings the permalinks won't be adjusted. This pull request uses another approach if setting is set true.
There might be a better way to adjust the URL but this should be very reliable.